+++ packages/components/inputs/select-utils/src/custom-styled-select-options/index.ts
@@ -1,3 +1,3 @@
 export { SELECT_DROPDOWN_OPTION_TYPES } from './constants';
 export * from './custom-styled-select-options';
-export * from './optionStyle-checkbox-components';
+export * from './options-style-checkbox-components';
